requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
pytest==6.2.4
pytest-django==4.4.0
djangorestframework==3.12.4
pytest-mock==3.6.1
python-dotenv==0.20.0
django-mock-queries==2.1.7
six==1.16.0
dj_database_url
requests~=2.27.1